+++ LocalAuth.android.js
@@ -1,15 +1,21 @@
 /**
- * Stub of LocalAuth for Android.
- *
+ * Mostly a copy of https://github.com/naoufal/react-native-touch-id
  * @providesModule LocalAuth
  * @flow
  */
-'use strict';
+'use strict'
 
-var warning = require('warning');
+import { createError } from './error'
+import Errors from './data/errors'
+
+const noTouchID = new Promise.reject(createError(Errors.RCTTouchIDNotSupported))
 
 module.exports = {
-  test: function test() {
-    warning('Not yet implemented for Android.')
+  hasTouchID() {
+    return noTouchID
+  },
+
+  authenticate(opts) {
+    return noTouchID
   }
 }
