+++ index.js
@@ -1,0 +1,6 @@
+
+import {
+  Platform
+} from 'react-native'
+
+module.exports = Platform.OS === 'android' ? require('./LocalAuth.android') : require('./LocalAuth.ios')
